1st try with freebsd & amd64

+++ ptlib/src/ptlib/unix/svcproc.cxx
@@ -392,7 +392,7 @@ void PSystemLog::Output(Level level, const char * cmsg)
 #elif defined(BE_THREADS)
       thread_id tid = ::find_thread(NULL);
 #else
-      unsigned tid = (unsigned) pthread_self();
+      unsigned long tid = (unsigned long) pthread_self();
 #endif
       *out << "ThreadID=0x"
            << setfill('0') << ::hex
@@ -942,7 +942,7 @@ void PServiceProcess::PXOnAsyncSignal(int sig)
 #elif defined(BE_THREADS)
   thread_id tid = ::find_thread(NULL);
 #else
-  unsigned tid = (unsigned) pthread_self();
+  unsigned long tid = (unsigned long) pthread_self();
 #endif
   PThread * thread_ptr = activeThreads.GetAt(tid);
